pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-mips-scheduler-env"
version = "0.1.0"
description = "MIPS Pipeline Instruction Scheduling Environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "mips_scheduler_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = [
    "mips_scheduler_env",
    "mips_scheduler_env.server",
]
[tool.setuptools.package-dir]
"mips_scheduler_env" = "."
"mips_scheduler_env.server" = "server"
